feat(design-system): wire the wave-2 vocabulary into the library registry
Single integration commit over the ten vocabulary branches. Nothing here
changes what a component looks like; it makes the promotions visible to
the registry, the index, and the gates.

Registry:
- StatusDot and IconTile join PRIMITIVES_ENTRIES; NoticeBanner, Card,
  Disclosure, RosterRow, SettingsPageBody, ProductPageShell join
  PATTERNS_ENTRIES; the tabs and panel kits spread their own entry
  modules in.
- PAGE_HEADER_ENTRY replaces the inline PageHeader row rather than
  adding beside it, so no subpath is declared twice.
- ProductPageShell's row moves down from the domain tier; ModelTable's
  row and demo are deleted with the component's move to its sole
  consumer; billingGateView is imported from its new presentation
  module.
- Pinned inventory 36/29/10/75 -> 38/38/10/86, confirmed against the
  parity test's physical globs rather than arithmetic alone.
- The two flat entry files (status-dot, roster-row) join the rest under
  library/entries/.

Kit cohesion: PaneOptionsMenuItem descends into patterns/panel/ beside
PanelHeaderEntry, with its four importers rewritten. PaneIconButton
stays put — it is a root primitive with consumers outside the right
panel, so moving it is a tier change, not a file move.

NoticeBanner drops the `success` tone: none of the 16 catalogued
duplicate sites use it, and a variant with no consumer is speculation.

Docs: index rows for every new component (each citing the duplicate
evidence that admitted it), the dead ProductPageShell/ModelTable links
repaired, the PrStatusBadge and BillingGateState rows re-truthed against
their presenter splits, and the Current Gaps entries for kit
directories, domain-tier grandfathering, and pending-promotion shapes
brought back in line. styling.md's Card Surfaces section now points at
the Card pattern and names bg-surface-elevated-secondary (bg-foreground/5
is rejected by the appearance gate) and bg-sidebar (nothing paints a rail
with bg-sidebar-background, and the dark values differ).
